pyproject.toml
[project]
name = "open-unlearning"
version = "0.1.0"
description = "A library for machine unlearning in LLMs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Vineeth Dorna" },
  { name = "Anmol Mekala" },
]
dependencies = [
  "huggingface-hub==0.36.0",
  "transformers==4.51.3",
  "hf-xet==1.2.0",
  "numpy==2.2.3",
  "hydra-core==1.3",
  "hydra_colorlog==1.2.0",
  "torch==2.4.1",
  "datasets==3.0.1",
  "accelerate==0.34.2",
  "bitsandbytes==0.44.1",
  "rouge-score==0.1.2",
  "scipy==1.14.1",
  "tensorboard==2.18.0",
  "scikit-learn==1.5.2",
  "deepspeed==0.15.4",
  "wandb==0.21.4",
  "pytest[dev]>=9.0.3",
  "flash-attn==2.6.3",
]

[project.optional-dependencies]
"lm-eval" = [
  "lm-eval==0.4.11",
]
dev = [
  "pre-commit==4.0.1",
  "ruff==0.6.9",
]

[project.urls]
Homepage = "https://github.com/locuslab/open-unlearning"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
